pyproject.toml
[tool.poetry]
name = "allauth-jinja"
version = "0.0.1"
description = ""
authors = ["Eddie Cohen <eddie@lightmatter.com>"]
readme = "README.md"
packages = [{include = "allauth_jinja"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
django-allauth = "^0.56.1"
django-jinja = "^2.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
